requirements.txt
python-gitlab
google-generativeai
Flask
python-dotenv
google-cloud-firestore